### 2019-10-24 #1835 @whedon
[REVIEW]: RepnDecomp: A GAP package for decomposing linear representations of finite groups
**Submitting author:** @kaashif (<a href="http://orcid.org/0000-0001-8581-5804">Kaashif Hymabaccus</a>)
**Repository:** <a href="https://github.com/gap-packages/RepnDecomp/" target ="_blank">https://github.com/gap-packages/RepnDecomp</a>
**Version:** 1.0.1
