### 2019-11-03 #1835 @falque
Hello there :)

I have started to review the submission, and since I am new to this whole reviewing process I am very very sorry if I miss some obvious elements, ask stupid questions, or such things...

The check list I have to fill has a section called Software paper, I assume it to be this: https://github.com/openjournals/joss-papers/blob/joss.01835/joss.01835/10.21105.joss.01835.pdf.
And I also assume that the items I need to check are directly related to this paper, but I have only found the information dwelled on by the State of the field item in the README file, so maybe it should be moved or copied to the paper ?
As for the authors, two names appear in the paper, but only one in the documentation, I guess it is because the other one will not be maintaining the code. Even though, in PackageInfo.g, Dmitrii Pasechnik is indicated as neither maintaining nor authoring, is that normal ?

Now, again, I assume the Documentation section of the checklist is about the documentation, which I reckon is that hosted at https://kaashif.gitlab.io/decomp/chap1.html#X789233A47A277072. Maybe I missed something, but I do not find there the Installation Instructions (second item of the list). I did find the list of dependencies in PackageInfo.g but maybe something explicit is expected in the documentation itself ?

These are of course details and I have not taken a deep look yet but for now it looks very nice !
