### 2019-11-03 #1835 @alex-konovalov
@falque
> Even though, in PackageInfo.g, Dmitrii Pasechnik is indicated as neither maintaining nor authoring, is that normal ?
Yes, one could now to this in PackageInfo.g in GAP, to give credit to those contributors who are neither authors nor maintainers.
