### 2019-11-16 #1835 @kaashif
@falque Very sorry for the delay in reply, I was on holiday for the last 10 days.

> Hello there :)
>
> I have started to review the submission, and since I am new to this whole reviewing process I am very very sorry if I miss some obvious elements, ask stupid questions, or such things...

I am also new to this process, there is no way your questions could be more stupid than mine :)

> The check list I have to fill has a section called Software paper, I assume it to be this: https://github.com/openjournals/joss-papers/blob/joss.01835/joss.01835/10.21105.joss.01835.pdf.
> And I also assume that the items I need to check are directly related to this paper, but I have only found the information dwelled on by the State of the field item in the README file, so maybe it should be moved or copied to the paper ?

Yes, I suppose this is useful context which belongs in the paper.

> Now, again, I assume the Documentation section of the checklist is about the documentation, which I reckon is that hosted at https://kaashif.gitlab.io/decomp/chap1.html#X789233A47A277072. Maybe I missed something, but I do not find there the Installation Instructions (second item of the list). I did find the list of dependencies in PackageInfo.g but maybe something explicit is expected in the documentation itself ?

Yes, I forgot installation instructions, I will add these.

> These are of course details and I have not taken a deep look yet but for now it looks very nice !

Thanks. I have a problem though, my employer still needs to approve any publications or contributions to software I make (this is a one-time thing). This should be a formality since this software is obviously (at least to me) unrelated to my work and not even in the same language, but these kinds of reviews are low priority and it takes a while for those responsible to actually do them.

I will chase some people on Monday and hopefully I'll make your suggested changes then.
