### 2019-11-22 #1835 @kaashif
A quick update: I contacted the correct people and they will get around to reviewing my request eventually. I hope I'm not being too cautious, but people do get fired over this kind of thing sometimes, especially when working on sensitive financial software. Sorry I can't move this along a bit more promptly.
