### 2019-11-27 #1835 @kaashif
The massive corporate bureacracy has churned and I have been told there are no IP concerns in this project! I'll make the changes and push them on the weekend. By that I mean the test fixes @alex-konovalov pointed out I need and the missing/jumbled docs @falque pointed out.

From now on, any delay is my fault :)
