### 2019-11-30 8f0b232
Add comment in paper on state of the field
