### 2019-11-30 4778480
Add installation instructions to README
