+++ public/chap0.txt
@@ -27,6 +27,9 @@
   
   1 [33X[0;0YIntroduction[133X
     1.1 [33X[0;0YGetting started with RepnDecomp[133X
+      1.1-1 [33X[0;0YInstallation[133X
+      1.1-2 [33X[0;0YNote on what is meant by a representation[133X
+      1.1-3 [33X[0;0YAPI Overview[133X
   2 [33X[0;0YIsomorphisms between representations[133X
     2.1 [33X[0;0YFinding explicit isomorphisms[133X
       2.1-1 LinearRepresentationIsomorphism
