+++ public/chap1.txt
@@ -8,9 +8,43 @@
   representation   [23X\rho  :  G  \to  GL(V)[123X  where  [23XG[123X  is  finite  and  [23XV[123X  is  a
   finite-dimensional [23X\mathbb{C}[123X-vector space.[133X
   
+  
+  [1X1.1-1 [33X[0;0YInstallation[133X[101X
+  
   [33X[0;0YTo  install  this  package,  refer  to  the installation instructions in the
   README    file    in    the    source    code.    It    is   located   here:
   [7Xhttps://gitlab.com/kaashif/decomp/blob/master/README.md[107X.[133X
+  
+  
+  [1X1.1-2 [33X[0;0YNote on what is meant by a representation[133X[101X
+  
+  [33X[0;0YThroughout   this  documentation,  mathematical  terminology  is  used  e.g.
+  representation.  It  is  clear  what  is meant mathematically, but it is not
+  entirely  clear  what is meant in terms of GAP types - what are you supposed
+  to  pass  in  when I say "pass in a representation". Occasionally I will not
+  even  mention what we are passing in and assume the reader knows that [3Xrho[103X or
+  [23X\rho[123X  refers  to a representation. A representation we can use is, in GAP, a
+  homomorphism  from  a finite group to a matrix group where all matrices have
+  coefficients  in  a  cyclotomic  field ([10XCyclotomics[110X is the union of all such
+  fields in GAP). You can check whether something you want to pass is suitable
+  with the function [2XIsFiniteGroupLinearRepresentation[102X ([14X4.1-1[114X).[133X
+  
+  [33X[0;0YHere's an example of a representation [3Xrho[103X in GAP:[133X
+  
+  [4X[32X  Example  [32X[104X
+    [4X[28X gap> G := SymmetricGroup(3);[128X[104X
+    [4X[28X Sym( [ 1 .. 3 ] )[128X[104X
+    [4X[28X gap> images := List(GeneratorsOfGroup(G), g -> PermutationMat(g, 3));[128X[104X
+    [4X[28X [ [ [ 0, 1, 0 ], [ 0, 0, 1 ], [ 1, 0, 0 ] ],[128X[104X
+    [4X[28X   [ [ 0, 1, 0 ], [ 1, 0, 0 ], [ 0, 0, 1 ] ] ][128X[104X
+    [4X[28X gap> rho := GroupHomomorphismByImages(G, Group(images));[128X[104X
+    [4X[28X [ (1,2,3), (1,2) ] -> [ [ [ 0, 1, 0 ], [ 0, 0, 1 ], [ 1, 0, 0 ] ],[128X[104X
+    [4X[28X   [ [ 0, 1, 0 ], [ 1, 0, 0 ], [ 0, 0, 1 ] ] ][128X[104X
+    [4X[28X [128X[104X
+  [4X[32X[104X
+  
+  
+  [1X1.1-3 [33X[0;0YAPI Overview[133X[101X
   
   [33X[0;0YThe  algorithms  implemented  can be divided into two groups: methods due to
   Serre  from  his  book Linear Representations of Finite Groups, and original
