### 2019-11-30 650d93e
Add notes about flaky tests and contributing
