### 2019-11-30 #1835 @kaashif
Changes I have made due to comments by @falque:

* Copied the comment about the state of the field to the paper.md. The comment is just that no open source software exists to do this but methods are well-known, so it's quite short.

* Added installation instructions to README and a link to the README in the first section of the documentation

From comments by @alex-konovalov:

* Fixed tests. The error he reported [here](https://travis-ci.org/gap-infra/gap-docker-pkg-tests-master-devel/jobs/584474552) was just a difference in how vector spaces were printed, before it was "0 generators" now it's "dimension 0". They do pass now on the newest GAP.

From really old comments by @dimpase:

* It was unclear what "representation" meant in terms of GAP (you need a finite group, coefficients in Cyclotomics (NOT finite fields)) and in some places I just use "rho" without even mentioning that it's a representation. I added a paragraph about this in the first manual section.

All changes have been pushed [here](https://gitlab.com/kaashif/decomp/).
