### 2019-12-09 #1835 @alex-konovalov
@kaashif @dimpase is there a plan to submit this package for the redistribution with GAP?
