### 2019-12-09 #1835 @dimpase
@alex-konovalov - will we make the coming GAP 4.11 release if we do it this week?
