### 2019-12-09 #1835 @alex-konovalov
If it will satisfy the requirements and pass integration tests, then yes. To go ahead, please suggest depositing the package by email to GAP Support, and submit a PR to https://github.com/gap-system/gap-distribution/blob/master/DistributionUpdate/PackageUpdate/currentPackageInfoURLList adding its PackageInfo.g URL.
