### 2019-12-09 #1835 @alex-konovalov
@dimpase however, you need to establish a proper package webpage and releases process first - at the moment, https://gitlab.com/kaashif/decomp/blob/master/PackageInfo.g just points to the GitLab repository, and the link to package archive will not work. Maybe you want to migrate to https://github.com/gap-packages first, and then use ReleaseTools and GitHubPagesForGAP to set them up using our standard tools? (I think I have suggested that to you elsewhere, but can't see in the discussion above).
