### 2019-12-14 #1835 @kaashif
Yeah I think migrating to https://github.com/gap-packages is a good idea, I've started doing that now. I'll also do the ReleaseTools stuff, change all of the URLs etc. I will also change the URL in the paper.md.
