### 2019-12-14 #1835 @kaashif
@VivianePons Is there a way for me to change the source repository URL in the JOSS paper to point to https://github.com/gap-packages/RepnDecomp, since this is the place where proper GAP packages live?
