### 2019-12-14 0d1cd7e
Remove generated docs (will be in gh-pages)
