+++ public/chap5.txt
@@ -1,132 +1,0 @@
-  
-  [1X5 [33X[0;0YComputing decompositions of representations[133X[101X
-  
-  
-  [1X5.1 [33X[0;0YBlock diagonalizing[133X[101X
-  
-  [33X[0;0YGiven  a  representation [23X\rho : G \to GL(V)[123X, it is often desirable to find a
-  basis  for [23XV[123X that block diagonalizes each [23X\rho(g)[123X with the block sizes being
-  as  small  as possible. This speeds up matrix algebra operations, since they
-  can now be done block-wise.[133X
-  
-  [1X5.1-1 BlockDiagonalBasisOfRepresentation[101X
-  
-  [33X[1;0Y[29X[2XBlockDiagonalBasisOfRepresentation[102X( [3Xrho[103X ) [32X function[133X
-  [6XReturns:[106X  [33X[0;10YBasis for [23XV[123X that block diagonalizes [23X\rho[123X.[133X
-  
-  [33X[0;0YLet  [23XG[123X  have  irreducible  representations  [23X\rho_i[123X,  with  dimension [23Xd_i[123X and
-  multiplicity [23Xm_i[123X. The basis returned by this operation gives each [23X\rho(g)[123X as
-  a block diagonal matrix which has [23Xm_i[123X blocks of size [23Xd_i \times d_i[123X for each
-  [23Xi[123X.[133X
-  
-  [1X5.1-2 BlockDiagonalRepresentation[101X
-  
-  [33X[1;0Y[29X[2XBlockDiagonalRepresentation[102X( [3Xrho[103X ) [32X function[133X
-  [6XReturns:[106X  [33X[0;10YRepresentation  of  [23XG[123X  isomorphic to [23X\rho[123X where the images [23X\rho(g)[123X
-            are block diagonalized.[133X
-  
-  [33X[0;0YThis      is      just     a     convenience     operation     that     uses
-  [2XBlockDiagonalBasisOfRepresentation[102X  ([14X5.1-1[114X)  to  calculate  the basis change
-  matrix and applies it to put [23X\rho[123X into the block diagonalised form.[133X
-  
-  
-  [1X5.2 [33X[0;0YAlgorithms due to the authors[133X[101X
-  
-  [1X5.2-1 REPN_ComputeUsingMyMethod[101X
-  
-  [33X[1;0Y[29X[2XREPN_ComputeUsingMyMethod[102X( [3Xrho[103X ) [32X attribute[133X
-  [6XReturns:[106X  [33X[0;10YA record in the same format as [2XREPN_ComputeUsingSerre[102X ([14X5.3-4[114X)[133X
-  
-  [33X[0;0YComputes  the same values as [2XREPN_ComputeUsingSerre[102X ([14X5.3-4[114X), taking the same
-  options.    The    heavy    lifting    of    this    method   is   done   by
-  [2XLinearRepresentationIsomorphism[102X   ([14X2.1-1[114X),  where  there  are  some  further
-  options that can be passed to influence algorithms used.[133X
-  
-  [1X5.2-2 REPN_ComputeUsingMyMethodCanonical[101X
-  
-  [33X[1;0Y[29X[2XREPN_ComputeUsingMyMethodCanonical[102X( [3Xrho[103X ) [32X attribute[133X
-  [6XReturns:[106X  [33X[0;10YA record in the same format as [2XREPN_ComputeUsingMyMethod[102X ([14X5.2-1[114X).[133X
-  
-  [33X[0;0YPerforms  the  same  computation  as  [2XREPN_ComputeUsingMyMethod[102X ([14X5.2-1[114X), but
-  first    splits   the   representation   into   canonical   summands   using
-  [2XCanonicalDecomposition[102X  ([14X5.3-1[114X).  This might reduce the size of the matrices
[…]
