+++ public/chap3.txt
@@ -1,36 +1,0 @@
-  
-  [1X3 [33X[0;0YAlgorithms for unitary representations[133X[101X
-  
-  
-  [1X3.1 [33X[0;0YUnitarising representations[133X[101X
-  
-  [1X3.1-1 UnitaryRepresentation[101X
-  
-  [33X[1;0Y[29X[2XUnitaryRepresentation[102X( [3Xrho[103X ) [32X function[133X
-  [6XReturns:[106X  [33X[0;10YA record with fields L and unitary_rep such that [3Xrho[103X is isomorphic
-            to unitary_rep, differing by a change of basis L.[133X
-  
-  [1X3.1-2 IsUnitaryRepresentation[101X
-  
-  [33X[1;0Y[29X[2XIsUnitaryRepresentation[102X( [3Xrho[103X ) [32X function[133X
-  [6XReturns:[106X  [33X[0;10YWhether  [3Xrho[103X  is  unitary,  i.e.  for  all [23Xg \in G[123X, [23X\rho(g^{-1}) =
-            \rho(g)^*[123X (where [23X^*[123X denotes the conjugate transpose).[133X
-  
-  [1X3.1-3 LDLDecomposition[101X
-  
-  [33X[1;0Y[29X[2XLDLDecomposition[102X( [3XA[103X ) [32X function[133X
-  [6XReturns:[106X  [33X[0;10Ya   record   with   two   fields,   L   and   D   such  that  [23XA  =
-            L\mbox{diag}(D)L^*[123X.  [23XD[123X  is  the  [23X1 \times n[123X vector which gives the
-            diagonal matrix [23X\mbox{diag}(D)[123X (where [3XA[103X is an [23Xn \times n[123X matrix).[133X
-  
-  
-  [1X3.2 [33X[0;0YDecomposing unitary representations[133X[101X
-  
-  [1X3.2-1 IrreducibleDecompositionDixon[101X
-  
-  [33X[1;0Y[29X[2XIrreducibleDecompositionDixon[102X( [3Xrho[103X ) [32X function[133X
-  [6XReturns:[106X  [33X[0;10Ya list of irreps in the decomposition of [3Xrho[103X[133X
-  
-  [33X[0;0YNOTE:  this  is not implemented yet. Assumes that [3Xrho[103X is unitary and uses an
-  algorithm due to Dixon to decompose it into unitary irreps.[133X
-  

[…]
