### 2019-12-14 0c2b5f1
Setup gh-pages based on GitHubPagesForGAP
