### 2019-12-14 c5b0b33
Remove shell script that just calls GAP script
