+++ makedoc.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-gap --nointeract makedoc.g
