+++ .travis.yml
@@ -1,0 +1,40 @@
+language: c
+env:
+  global:
+    # For details how this works, and what other configurations you
+    # can make, please look at <https://github.com/gap-system/pkg-ci-scripts>
+    #- GAP_PKGS_TO_CLONE="SomeOtherPackage"
+    #- GAP_PKGS_TO_BUILD="io profiling SomeOtherPackage" # note: io and profiling always must be built
+
+addons:
+  apt_packages:
+    - libgmp-dev
+    - libreadline-dev
+    - zlib1g-dev
+
+matrix:
+  include:
+    - env: GAPBRANCH=master
+    - env: GAPBRANCH=stable-4.10
+    - env: GAPBRANCH=stable-4.9
+    - env: GAPBRANCH=master ABI=32
+      addons:
+        apt_packages:
+          - libgmp-dev:i386
+          - libreadline-dev:i386
+          - zlib1g-dev:i386
+          - gcc-multilib
+          - g++-multilib
+
+branches:
+  only:
+    - master
+
+before_script:
+  - git clone https://github.com/gap-system/pkg-ci-scripts.git scripts
+  - scripts/build_gap.sh
+script:
+  - scripts/build_pkg.sh && scripts/run_tests.sh
+after_script:
+  - bash scripts/gather-coverage.sh
+  - bash <(curl -s https://codecov.io/bash)
