### 2019-12-14 714bb94
Remove gap 4.9 from travis (not supported)
