+++ .travis.yml
@@ -16,7 +16,6 @@ matrix:
   include:
     - env: GAPBRANCH=master
     - env: GAPBRANCH=stable-4.10
-    - env: GAPBRANCH=stable-4.9
     - env: GAPBRANCH=master ABI=32
       addons:
         apt_packages:
