### 2019-12-14 2e3f1af
Add CI badge
