### 2019-12-14 7754401
Remove extra travis environments
