### 2019-12-19 #1835 @VivianePons
I have updated the github issue. @openjournals/dev can you make the requied change wherever it is needed? Thanks!
