### 2019-12-20 #1835 @arfon
> I have updated the github issue. @openjournals/dev can you make the requied change wherever it is needed? Thanks!

Done.
