### 2019-12-22 1095d44
Update README so it's about the package
