### 2020-01-05 b1f5647
Convert examples to tests in travis file
