+++ .travis.yml
@@ -24,6 +24,7 @@ before_script:
   - git clone https://github.com/gap-system/pkg-ci-scripts.git scripts
   - scripts/build_gap.sh
 script:
+  - ./makedoc.sh
   - scripts/build_pkg.sh && scripts/run_tests.sh
 after_script:
   - bash scripts/gather-coverage.sh
