### 2020-01-20 #1835 @alex-konovalov
Thanks for the reminder @VivianePons - I still have to evaluate the new version and check how this is aligned with getting the package redistributed with GAP. I shall find time for this in February.
