### 2020-01-24 #1835 @kaashif
I'll have some time to work on this project this weekend. I think I can have all of the issues closed in a few days, I just didn't have much time over Christmas and New Year.
