### 2020-01-26 2464f07
Switch to standard Travis CI and Codecov setup for GAP packages
