### 2020-02-10 #1835 @alex-konovalov
@danielskatz @VivianePons the package has been submitted for the redistribution with GAP, testing outcomes were reported at https://github.com/gap-system/gap-distribution/pull/101. We are waiting now for the next release of the package to be published, then I will check its compatibility with GAP and other packages, and hopefully it will be suitable for the redisribituion with GAP. This will allow me to complete the review.
