+++ doc/chapInd.txt
@@ -4,7 +4,6 @@
   
   [2XAreRepsIsomorphic[102X 2.2-1 
   [2XBlockDiagonalBasisOfRepresentation[102X 5.1-1 
-  [2XBlockDiagonalMatrix[102X 4.4-1 
   [2XBlockDiagonalRepresentation[102X 5.1-2 
   [2XCanonicalDecomposition[102X 5.3-1 
   [2XCentralizerBlocksOfRepresentation[102X 6.1-1 
@@ -12,7 +11,7 @@
   [2XCharacterOfTensorProductOfRepresentations[102X 4.3-5 
   [2XClassSumCentralizer[102X 6.2-1 
   [2XClassSumCentralizerNC[102X 6.2-2 
-  [2XComposeHomFunction[102X 4.4-2 
+  [2XComposeHomFunction[102X 4.4-1 
   [2XDegreeOfRepresentation[102X 4.5-3 
   [2XDirectSumOfRepresentations[102X 4.5-2 
   [2XGroupSumBSGS[102X 4.2-1 
@@ -31,7 +30,6 @@
   [2XLinearRepresentationIsomorphism[102X 2.1-1 
   [2XLinearRepresentationIsomorphismSlow[102X 2.1-2 
   [2XPermToLinearRep[102X 4.5-4 
-  [2XReplicate[102X 4.4-3 
   [2XREPN_ComputeUsingMyMethod[102X, for IsGroupHomomorphism 5.2-1 
   [2XREPN_ComputeUsingMyMethodCanonical[102X, for IsGroupHomomorphism 5.2-2 
   [2XREPN_ComputeUsingSerre[102X, for IsGroupHomomorphism 5.3-4 
