+++ README.md
@@ -1,25 +1,32 @@
+[![Build Status](https://travis-ci.org/gap-packages/RepnDecomp.svg?branch=master)](https://travis-ci.org/gap-packages/RepnDecomp)
+[![Code Coverage](https://codecov.io/github/gap-packages/RepnDecomp/coverage.svg?branch=master&token=)](https://codecov.io/gh/gap-packages/RepnDecomp)
+
 # Computing decompositions of representations
 
-## Project Description
+## Overview
 
-This project aims to investigate and record algorithmic details and to
-produce a program (using GAP and/or Sagemath) for computing the
-decomposition of a representation ρ, of a finite group G over the
-complex numbers into irreducibles, as well as the corresponding
-decomposition of the centraliser of R. Currently, while methods for
-doing this are known (cf.  Serre's book "Linear Representations of
-Finite Groups"), there are no open-source computer programs that
-implement these methods, nor are details on how to achieve good
-performance of such an implementation published.
+The main function of this package is to compute the decomposition of a
+representation ρ, of a finite group G over the complex numbers into
+irreducibles, as well as the corresponding decomposition of the
+centraliser of R.
 
-This program will be useful in, for example, semidefinite programming
-and optimisation/feasibility problems involving coding theory, graph
+While methods for doing this were well known before this package was
+written (cf.  Serre's book "Linear Representations of Finite Groups"),
+there were no open-source computer programs that implemented these
+methods, nor were details on how to achieve good performance of such
+an implementation published.
+
+This package is useful in, for example, semidefinite programming and
+optimisation/feasibility problems involving coding theory, graph
 theory, algebraic geometry, combinatorics and more (see
 https://arxiv.org/abs/1007.2905 for more examples of possible
 applications). Specifically, it allows to achieve substantial
 reductions in the dimension of these problems; potentially known
 results, e.g. on upper bounds on sizes of nonlinear codes, could be
-improved with the help of the program.
+improved with the functions implemented in this package.
+
+This package was written as part of my (Kaashif Hymabaccus's) Master's
+degree at the University of Oxford, supervised by Dmitrii Pasechnik.
 
 ## Installation
 
@@ -41,7 +48,7 @@ functions provided:
 ```
 $ gap
 <some output>
-gap> gap> LoadPackage("RepnDecomp");
+gap> LoadPackage("RepnDecomp");
 ───────────────────────────────────────────────────────────────────────────────
 Loading  GRAPE 4.8.2 (GRaph Algorithms using PErmutation groups)
 by Leonard H. Soicher (http://www.maths.qmul.ac.uk/~lsoicher/).
@@ -69,6 +76,10 @@ Where `LDLDecomposition` is a function provided by this package.
 
 ## Testing
 
+First, generate the tests from the documentation:
+
+    $ gap -q < makedoc.g
+
 To run the tests, make sure you have all needed packages installed
 (GRAPE and also IO if you want to compute in parallel).
 
@@ -78,10 +89,8 @@ This will run all tests and (hopefully) pass. In order for the tests
 to be as useful as possible to me i.e. catch as many bugs as possible,
 there is a lot of randomness in them. This means that sometimes the
 tests pick a pathologically bad example to decompose which causes the
-tests to hang forever.
-
-The best way to fix it would be to send a PR to make the performance
-better :)
+tests to hang forever. The tests will be fixed to be more
+deterministic.
 
 ## Documentation
 
@@ -91,9 +100,12 @@ is generated from the source files and comments in the `lib/`
 directory of this repo, so you can also look there for the same
 information.
 
-There are also some examples in the `examples` directory, but the most
-complete set of examples are in the `tst` directory - the unit tests
-themselves.
[…]
