### 2020-02-15 #1835 @kaashif
Done, the manual examples, new non-flaky tests, and some changes from @alex-konovalov and @dimpase are now published in version 1.1.0: https://gap-packages.github.io/RepnDecomp/. Let me know what you think.
