### 2020-02-25 #1835 @VivianePons
Thank you @falque and thanks for the suggestion, I agree with you that the clearer the better.

Now I'm only waiting for @alex-konovalov , where are we in terms of GAP realeases and package testing?
