### 2020-02-25 #1835 @kaashif
>
>
> As far as I am concerned, everything is okay (installation, tests, documentation, well, you can see the checklist) ! As a tiny remark, in the Testing section of the README file, I would add at the beginning that this is all supposed to take place in the previously unzipped ~/.gap/pkg/RepnDecomp . To be sure, it is highly guessable, but it took me a short while to understand so I think it could only be an improvement, especially for non-experts.

Thanks for the suggestion, I'll change that now.
