### 2020-02-25 bfc7b3a
Improve test instructions
