### 2020-03-05 #1835 @alex-konovalov
@VivianePons some progress - RepnDecomp will be definitely included in GAP 4.11.0, which will be publicly announced shortly. I will finish the review after that. Hope we can get it out in March!
