### 2020-04-13 #1835 @alex-konovalov
@arfon thanks for the reminder. We published GAP 4.11.0 which includes RepnDecomp (https://www.gap-system.org/Releases/4.11.0.html), I will try to complete the review shortly.
