### 2020-04-21 #1835 @alex-konovalov
@arfon @kaashif It all looks good to me - some minor remarks:

- the package has been migrated to https://github.com/gap-packages/RepnDecomp so I did not tick the first box which refers to the old repository. The link in the paper needs to be updated.

- installation instructions at https://github.com/gap-packages/RepnDecomp/blob/master/README.md describe how to install a development version of the package from its repository, in `./gap/pkg` directory where users can keep their customisations. First, you now have a proper versioning, so I would be suggesting to download latest official release from https://gap-packages.github.io/RepnDecomp/ (unless the user wants the development version). Second, since the package is now redistributed with GAP, you can mention that users having GAP 4.11.0 or newer will not have to install the package, since it is already included in the GAP distribution).

- some wording in the paper needs to be improved, for example

```
When studying problems in algebra and combinatorics, it is
often useful to pass to a representation to more easily understand the
structure of the problem.
```
to pass what? Also, "more easily" -> "easier" or "better"

```
This is also useful for computation, since
we can use representations to transform our algebraic problems into
linear algebra.
```
"into linear algebra" does not sound right to me.

```
finding a basis which block diagonalises the matrices of the
representation optimally,
```
Perhaps"which finds an optimal block diagonal form"...

```
* Decompose a representation into irreducibles
```
"into irreducible representations" ?

```
Our methods are not specific to cyclotomic numbers, but the GAP
algebra system ...
```
Either just "the GAP system" or "the GAP computational algebra system".
In fact, this is the first time GAP mentioned here. It would be useful to
add a reference ([see how](https://www.gap-system.org/Contacts/cite.html)).
