### 2020-04-21 #1835 @kaashif
Thanks for the feedback, I agree some of the phrasing is a bit awkward and I'll address your comments shortly. Also thanks for reminding me about the new installation instructions now that there are real releases and versions.

One comment:

> "into linear algebra" does not sound right to me.

It sounds fine to me from a linguistic perspective, but maybe this isn't how mathematicians really speak. Maybe I should say "representations allow us to use tools from linear algebra to solve problems in group theory" or something along those lines.
