### 2020-04-21 #1835 @alex-konovalov
@kaashif your new formulation sounds perfect to me, thanks! I think, saying "into linear algebra" is acceptable informally, but not in a paper.
