### 2020-04-22 #1835 @dimpase
Thanks a lot for looking into this.

Just one remark - I think "irredicubles" is an acceptable domain terminology - the rest I agree with.
