### 2020-04-22 #1835 @alex-konovalov
Informally, it's fine, and it's clear from the context whether one means irreducible polynomials, or characters, or representations - but in the paper, I'd still prefer to avoid using adjectives as nouns.
