### 2020-04-25 eba2ba7
Apply suggestions from @alex-konovalov
