+++ paper.bib
@@ -48,3 +48,12 @@ MRREVIEWER = {W. Feit},
  volume = {24},
  year = {1970}
 }
+
+@manual{gap:2020,
+    key          = "GAP",
+    organization = "The GAP~Group",
+    title        = "{GAP -- Groups, Algorithms, and Programming,
+                    Version 4.11.0}",
+    year         = 2020,
+    url          = "\url{https://www.gap-system.org}",
+}
