### 2020-04-25 #1835 @dimpase
text says "we produced a package for GAP" - the package should be named (and perhaps hyperlinked to its URL)
