### 2020-04-25 f511456
Improve installation instructions
