+++ paper.bib
@@ -55,5 +55,5 @@ MRREVIEWER = {W. Feit},
     title        = "{GAP -- Groups, Algorithms, and Programming,
                     Version 4.11.0}",
     year         = 2020,
-    url          = "\url{https://www.gap-system.org}",
+    url          = "https://www.gap-system.org",
 }
