### 2020-04-25 #1835 @kaashif
Suggestions from @alex-konovalov:

>     * the package has been migrated to https://github.com/gap-packages/RepnDecomp so I did not tick the first box which refers to the old repository. The link in the paper needs to be updated.

The latest PDF which I have generated contains a correct link to the repository at
https://github.com/gap-packages/RepnDecomp.

>     * installation instructions at https://github.com/gap-packages/RepnDecomp/blob/master/README.md describe how to install a development version of the package from its repository, in `./gap/pkg` directory where users can keep their customisations. First, you now have a proper versioning, so I would suggest to a user to download the latest official release from https://gap-packages.github.io/RepnDecomp/ (unless the user wants the development version). Second, since the package is now redistributed with GAP, you can mention that users having GAP 4.11.0 or newer will not have to install the package, since it is already included in the GAP distribution.

I improved the installation instructions in [the README](https://github.com/gap-packages/RepnDecomp#installation).

>     * some wording in the paper needs to be improved, for example
>
>
> ```
> When studying problems in algebra and combinatorics, it is
> often useful to pass to a representation to more easily understand the
> structure of the problem.
> ```
>
> to pass what? Also, "more easily" -> "easier" or "better"

I changed this to "When studying problems in algebra and combinatorics, it is often useful to also study associated representations to better understand the structure of the problem."

> ```
> This is also useful for computation, since
> we can use representations to transform our algebraic problems into
> linear algebra.
> ```
>
> "into linear algebra" does not sound right to me.

I changed this to "This is also useful for computation, since representations allow us to use tools from linear algebra to solve problems in group theory."

> ```
> finding a basis which block diagonalises the matrices of the
> representation optimally,
> ```
>
> Perhaps"which finds an optimal block diagonal form"...

I changed this to "finding a basis which produces an optimal block diagonal form of the representation". I think it is important to make it clear that the package is mainly for finding the basis. Finding an optimal block diagonal form is easy, involving only some character theory and use of off-the-shelf GAP functions, finding the basis is the tricky part.

> ```
> * Decompose a representation into irreducibles
> ```
>
> "into irreducible representations" ?

I changed this to "Decompose a representation into a direct sum of irreducible representations".

> ```
> Our methods are not specific to cyclotomic numbers, but the GAP
> algebra system ...
> ```
>
> Either just "the GAP system" or "the GAP computational algebra system".
> In fact, this is the first time GAP mentioned here. It would be useful to
> add a reference ([see how](https://www.gap-system.org/Contacts/cite.html)).

I changed this to "the GAP system" and I also added a citation for GAP before the bullet points.

From @dimpase:

I added a link to the package in the text of the paper, just before the bullet points. Also I fixed the URLs :)

From me:

I noticed that the unicode theta (as in Lovasz theta number) was not rendering in the PDFs, so I changed it to the TeX `\vartheta`, it looks right now.
