### 2020-04-25 #1835 @dimpase
OK, few more things:

- add a reference to Kaashif's MSc thesis (the implementation of the package was a part of KH's MSc thesis)
- add an acknowledgement to CODIMA (K.H. thanks Collaborative Computational Project (CCP) in the area of Computational Discrete Mathematics, supported by the EPSRC (EP/M022641/1) for supporting his participation in a GAP workshop)
- add an acknowledgement of OpenDreamKit (D.P. is partly funded by the OpenDreamKit Horizon 2020 European Research Infrastructures project (#676541))
