### 2020-04-25 4f8f7c1
Add master's thesis citation and acks
