+++ paper.bib
@@ -21,6 +21,14 @@
     pages = {293--320}
 }
 
+@ARTICLE{hymabaccus:2019,
+    author = {Hymabaccus, Kaashif},
+    title = {Decomposing Linear Representations of Finite Groups},
+    journal = {Unpublished master's thesis},
+    organization = {University of Oxford},
+    year = {2019},
+}
+
 @book {serre:1977,
     AUTHOR = {Serre, Jean-Pierre},
      TITLE = {Linear representations of finite groups},
