+++ paper.bib
@@ -25,6 +25,7 @@
     author = {Hymabaccus, Kaashif},
     title = {Decomposing Linear Representations of Finite Groups},
     journal = {Unpublished master's thesis},
+    publisher = {University of Oxford},
     organization = {University of Oxford},
     year = {2019},
 }
