### 2020-06-07 #1835 @kaashif
Sure. I made a new release and put it on Zenodo here: https://doi.org/10.5281/zenodo.3884306
