### 2020-06-09 59e103d
Update paper.bib
Adds DOIs
