+++ paper.bib
@@ -18,7 +18,8 @@
     title = {Group symmetry in interior-point methods for semidefinite programming},
     journal = {Optimization and Engineering},
     year = {1970},
-    pages = {293--320}
+    pages = {293--320},
+    doi = {10.1023/A:1015366416311}
 }
 
 @ARTICLE{hymabaccus:2019,
@@ -46,7 +47,7 @@ MRREVIEWER = {W. Feit},
 
 @article{dixon:1970,
  ISSN = {00255718, 10886842},
- URL = {http://www.jstor.org/stable/2004848},
+ doi = {10.2307/2004848},
  abstract = {How can you find a complete set of inequivalent irreducible (ordinary) representations of a finite group? The theory is classical but, except when the group was very small or had a rather special structure, the actual computations were prohibitive before the advent of high-speed computers; and there remain practical difficulties even for groups of relatively small orders (≤ 100). The present paper describes three techniques to help solve this problem. These are: the reduction of a reducible unitary representation into its irreducible components; the construction of a complete set of irreducible components; the construction of a complete set of irreducible unitary representations from a single faithful representation; and the calculation of the precise values of a group character from values which have only been computed approximately.},
  author = {John D. Dixon},
  journal = {Mathematics of Computation},
