### 2020-06-09 #1835 @kyleniemeyer
Hello @kaashif, in my final review of the article I noticed there were some DOIs missing (though our checker didn't catch them). I just submitted a small PR that adds these, can you merge it? https://github.com/gap-packages/RepnDecomp/pull/8
